# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Containr API
  description: |
    Self-hosted PaaS on Proxmox - Railway-like platform for container orchestration and deployment.
    
    ## Overview
    Containr provides a comprehensive platform for deploying, managing, and scaling containerized applications on Proxmox infrastructure. This API allows you to manage projects, services, deployments, databases, and more through a RESTful interface.
    
    ## Authentication
    The API uses JWT-based authentication. Include your JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Base URL
    ```
    https://your-containr-domain.com/api/v1
    ```
  version: 1.0.0
  contact:
    name: Containr Documentation
    url: https://github.com/your-org/containr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-containr-domain.com/api/v1
    description: Production server
paths:
...
  /services/{id}/deployments:
    get:
      tags:
        - Deployments
      summary: List deployments
      description: Retrieve recent deployments for a specific service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '200':
          description: Deployments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentListResponse'

    post:
      tags:
        - Deployments
      summary: Create deployment
      description: Trigger a new deployment for a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '201':
          description: Deployment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
  /deployments:
    get:
      tags:
        - Deployments
      summary: List recent deployments
      description: Latest deployments across all services owned by the caller — feeds the dashboard deploy feed
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Recent deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        service_id:
                          type: string
                        service_name:
                          type: string
                        project_name:
                          type: string
                        status:
                          type: string
                        image_name:
                          type: string
                        started_at:
                          type: string
                          nullable: true
                        completed_at:
                          type: string
                          nullable: true
                        created_at:
                          type: string
  /deployments/{id}:
    get:
      tags:
        - Deployments
      summary: Get deployment
      description: Retrieve a deployment with build/runtime details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
      responses:
        '200':
          description: Deployment retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment:
                    $ref: '#/components/schemas/Deployment'
...
  /templates/{id}/deploy:
    post:
      tags:
        - Templates
      summary: Create resource from template
      description: |
        Create a new resource in a project using a selected template.
        For `database` templates, this creates a managed database service.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployTemplateRequest'
      responses:
        '201':
          description: Service created from template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployTemplateResponse'

  # Agent Onboarding Tokens
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Deployment:
      type: object
      properties:
        id:
          type: string
          description: Deployment ID
        service_id:
          type: string
          description: Service ID
        commit_hash:
          type: string
          description: Source commit hash
        status:
          type: string
          description: Deployment status
        image_name:
          type: string
          description: Deployed image name
        image_tag:
          type: string
          description: Deployed image tag
        build_log:
          type: string
          description: Build log content
        runtime_log:
          type: string
          description: Runtime log content
        error:
          type: string
          description: Deployment error
        started_at:
          type: string
          format: date-time
          description: Deployment start timestamp
        completed_at:
          type: string
          format: date-time
          description: Deployment completion timestamp
        created_at:
          type: string
          format: date-time
          description: Deployment creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Deployment update timestamp
    DeploymentListResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/Deployment'
    CreateDeploymentRequest:
      type: object
      properties:
        commit_hash:
          type: string
          description: Source commit hash
        branch:
          type: string
          description: Source branch
        trigger:
          type: string
          description: Trigger source (manual, restart, webhook)
        env_vars:
          type: object
          additionalProperties:
            type: string
          description: Runtime environment variable overrides
...
    DeployTemplateRequest:
      type: object
      required:
        - project_id
        - name
      properties:
        project_id:
          type: string
          description: Target project ID
        name:
          type: string
          description: Name for the created service
        plan:
          type: string
          enum: [hobby, starter, standard, business]
          description: Optional managed database plan for database templates
        region:
          type: string
          description: Optional managed database region for database templates
        variables:
          type: object
          additionalProperties:
            type: string
          description: Variable overrides for template deployment
    DeployTemplateResponse:
      type: object
      properties:
        resource:
          type: string
          enum: [service, database]
          description: Type of resource created by template deployment
        service_id:
          type: string
          description: Newly created service ID (set for non-database templates)
        database_id:
          type: string
          description: Newly created managed database ID (set for database templates)
        message:
          type: string
          description: Deployment result message
...